% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceWorker.R
\name{addJS__}
\alias{addJS__}
\title{Add JavaScript}
\usage{
addJS__(filename)
}
\arguments{
\item{filename}{\itemize{
\item the desired JS file
}}
}
\value{
The JS file
}
\description{
Fetches JavaScript file with its path.
}
\examples{

\dontRun{

}

}
\keyword{internal}
