% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{createIconsRefs__}
\alias{createIconsRefs__}
\title{Create Icons References}
\usage{
createIconsRefs__()
}
\value{
An array of containing the icons
}
\description{
Creates an array containing all the icons to be loaded into the PWA mainifest
file.
}
\examples{

\dontrun {

}

}
\keyword{internal}
