% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{createManifest__}
\alias{createManifest__}
\title{Create Manifest}
\usage{
createManifest__(config)
}
\arguments{
\item{config}{\itemize{
\item a list containing:
}

\itemize{
\item{name}{The app name}
\item{shortname}{The app shortname}
\item{display}{The display value}
\item{backgroundcolor}{The background color}
\item{themecolor}{The theme color}
\item{orientation}{The orientation}
\item{hasIcons}{Whether or not the package has icons to be added}
\item{version}{The peckage version}
}}
}
\value{
The needed 'manifest.json' to be placed in the HMTL head
}
\description{
Generates the required 'manifest.json'
}
\examples{

\dontrun{
    createManifest(list(
        name = 'testing',
        shorname = 'test',
        display = 'standalone',
        backgroundcolor = '#fdfdfd',
        themecolor = '#db4938',
        orientation = 'portrait-primary'
    ))
}

}
\keyword{internal}
