% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceWorker.R
\name{createServiceWorker__}
\alias{createServiceWorker__}
\title{Create Service Worker}
\usage{
createServiceWorker__(websiteVersion)
}
\arguments{
\item{websiteVersion}{\itemize{
\item The current verion
}}
}
\value{
Whehter or not the operation has been successful completed
}
\description{
Generates the required 'serviceWroker.js'
}
\examples{

\dontRun{

}

}
\keyword{internal}
