% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{fetchIcons__}
\alias{fetchIcons__}
\title{Fetch Icons}
\usage{
fetchIcons__(icons)
}
\arguments{
\item{icons}{\itemize{
\item the icon folder
}}
}
\value{
The icons to fetch its metadata
}
\description{
Looks for files in the 'www/icons/' folder.
}
\examples{

\dontrun{
    fetchIcons__()
}

}
\keyword{internal}
