% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{inferIconMetadata__}
\alias{inferIconMetadata__}
\title{Infer Icon Metadata}
\usage{
inferIconMetadata__(icon)
}
\arguments{
\item{icon}{\itemize{
\item icon filename
}}
}
\value{
the desired metadata
}
\description{
Infer the incon metadata
}
\examples{

\dontrun {

}

}
