% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPWA.R
\name{shinyPWA}
\alias{shinyPWA}
\title{Shiny PWA}
\usage{
shinyPWA(config)
}
\arguments{
\item{config}{\itemize{
\item a list containing:
}

\itemize{
\item{name}{The app name}
\item{shortname}{The app shortname}
\item{display}{The display value}
\item{backgroundcolor}{The background color}
\item{themecolor}{The theme color}
\item{orientation}{The orientation}
\item{hasIcons}{Whether or not the package has icons to be added}
\item{version}{The peckage version}
}}
}
\value{
The needed 'manifest.json' to be placed in the HMTL head
}
\description{
Creates the files needed to transform your Shiny application in a Progressive
Web App (PWA).
}
\examples{

\dontrun{
    shinyPWA(list(
        name = 'testing',
        shorname = 'test',
        display = 'standalone',
        backgroundcolor = '#fdfdfd',
        themecolor = '#db4938',
        orientation = 'portrait-primary'
    ))
}

}
