% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPWA.R
\name{shinyPWA__}
\alias{shinyPWA__}
\title{Shiny PWA - input validator}
\usage{
shinyPWA__(config)
}
\arguments{
\item{config}{\itemize{
\item a list containing:
}

\itemize{
\item{name}{The app name}
\item{shortname}{The app shortname}
\item{display}{The display value}
\item{backgroundcolor}{The background color}
\item{themecolor}{The theme color}
\item{orientation}{The orientation}
\item{hasIcons}{Whether or not the package has icons to be added}
\item{version}{The peckage version}
}}
}
\value{
Whether or not the input is valid
}
\description{
Checks whether or not the input provided trought the user is a correct one.
}
\examples{

\dontrun{
    shinyPWA__(list(
        name = 'testing',
        shorname = 'test',
        display = 'standalone',
        backgroundcolor = '#fdfdfd',
        themecolor = '#db4938',
        orientation = 'portrait-primary'
    ))
}

}
\keyword{internal}
