% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{createManifest__}
\alias{createManifest__}
\title{Create Manifest}
\usage{
createManifest__()
}
\value{
Whether or not the operation has been successful
}
\description{
Generates the required 'manifest.json'
}
\keyword{internal}
