% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{createManifestJSON__}
\alias{createManifestJSON__}
\title{Create Manifest JSON}
\usage{
createManifestJSON__(name, shortname, starturl, display)
}
\description{
Creates the 'manifest.json' file
}
\examples{

\dontRun{

}

}
\keyword{internal}
