% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{fetchIcons__}
\alias{fetchIcons__}
\title{Fetch Icons}
\usage{
fetchIcons__()
}
\value{
the name of all the icons presented in the 'icon' folder
}
\description{
Looks for files in the 'www/icons/' folder.
}
\examples{

\dontRun {

}

}
\keyword{internal}
