% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPWA.R
\name{shinyPWA}
\alias{shinyPWA}
\title{Shiny PWA}
\usage{
shinyPWA()
}
\value{
Whether or the operation has been successful completed
}
\description{
Creates the files needed to transform your Shiny application in a Progressive
Web App (PWA).
}
\examples{

\dontrun{

}

}
