% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPWA.R
\name{shinyPWA__}
\alias{shinyPWA__}
\title{Shiny PWA - input validator}
\usage{
shinyPWA__()
}
\arguments{
\item{config}{\itemize{
\item a list containing:
}

\itemize{
\item name
\item shortname
\item display
\item backgroundcolor
\item themecolor
\item orientation
\item hasIcons
\item version
}}
}
\value{
Whether or not the input is valid
}
\description{
Checks whether or not the input provided trought the user is a correct one.
}
\examples{

\dontrun{

}

}
\keyword{internal}
